third_party/WebKit/LayoutTests/fast/forms/select/exceptions.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("This test should trigger the single exception on HTMLSelectElement, and verify that the message is reasonably helpful.");

        var el = document.createElement('select');

        shouldThrow("el[1] = 0;");
        shouldThrow("el.namedItem()");
    </script>
</body>
</html>
