third_party/blink/web_tests/fast/forms/select/exceptions-expected.txt

This test should trigger the single exception on HTMLSelectElement, and verify that the message is reasonably helpful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS el[1] = 0; threw exception TypeError: Failed to set an indexed property [1] on 'HTMLSelectElement': parameter 2 is not of type 'HTMLOptionElement'..
PASS el.namedItem() threw exception TypeError: Failed to execute 'namedItem' on 'HTMLSelectElement': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
